refactor(infrastructure) : RacingGameServiceImpl 패키지 분리 및 리팽토링
- model 패키지에서는 View 를 출력하는 로직이 있으면 패키지 복잡성을 높일 수 있으므로 분리
- RacingGameServiceImpl.java 를 infrastructure 패키지로 이동
- 개행추가 코드 제거 및 UserOutput 에 개행처리 코드 추가
